# Optimize host calls in the wasmi interpreter