This commit does two things to speed-up host calls:

- Reuse heap allocations for consecutive host calls to hold arguments.
- More efficient copying of stack values to host argument buffer.

## Benchmarks

No noticeable performance impact was measured, however, that's mainly because the current set of benchmarks do not cover the particular scenario of performing lots of host function calls.

### Before

```
test bench_regex_redux ... bench:   1,831,339 ns/iter (+/- 153,861)
test bench_rev_comp    ... bench:   1,661,062 ns/iter (+/- 146,230)
test bench_tiny_keccak ... bench:   1,276,610 ns/iter (+/- 132,734)
test fac_opt           ... bench:      23,268 ns/iter (+/- 4,547)
test fac_recursive     ... bench:      24,568 ns/iter (+/- 1,908)
test recursive_ok      ... bench:     615,655 ns/iter (+/- 43,826)
test recursive_trap    ... bench:      79,457 ns/iter (+/- 4,446)
```

### After

```
test bench_regex_redux ... bench:   1,802,265 ns/iter (+/- 130,377)
test bench_rev_comp    ... bench:   1,656,515 ns/iter (+/- 126,081)
test bench_tiny_keccak ... bench:   1,260,390 ns/iter (+/- 118,418)
test fac_opt           ... bench:      22,251 ns/iter (+/- 1,458)
test fac_recursive     ... bench:      23,990 ns/iter (+/- 2,073)
test recursive_ok      ... bench:     593,989 ns/iter (+/- 86,742)
test recursive_trap    ... bench:      78,235 ns/iter (+/- 5,871)
```