@Robbepop 2021-12-30
Added benchmarks in https://github.com/paritytech/wasmi/pull/291/commits/1455e9a3c1d0e4a5e2a9e2f26901107daaa05328.

Run the benchmark using:
```
cargo bench --manifest-path benches/Cargo.toml host_call
```

## Before

```
running 1 test
test host_calls        ... bench:     135,455 ns/iter (+/- 19,820)
```

## After

```
running 1 test
test host_calls        ... bench:      95,340 ns/iter (+/- 13,254)
```

So we can measure a speed-up of roughly 42% for the benchmark.